refactor(middleware): Create dir, make file for each
