+++ lib/populi_api/middleware/indifferent_hashify.rb
